fix: refactor notifications.js
